requirements.txt
joblib
keras
matplotlib
numpy
opencv-python
pandas
pyarrow
PyYAML
scikit-learn
seaborn
tables
tensorflow[and-cuda]
